requirements.txt
Flask==3.0.2
gunicorn==20.1.0
google-cloud-storage==2.18.2
google-generativeai==0.5.0
